+++ tests/testthat/reporters/error-setup.R
@@ -1,4 +1,6 @@
 f <- function() g()
 g <- function() h()
 h <- function() stop("!")
+
+local_edition(2)
 setup(f())